+++ __tests__/deposit.service.test.ts
@@ -1,5 +1,5 @@
 import { Deposit, PrismaClient } from "@prisma/client";
-import CONFIG from "@/config";
+import { COIN_TRANSFER_STATUS, DEPOSIT_STATUS } from "@/config";
 import { DepositServices } from "@/components/deposits/services";
 
 let service: DepositServices;
@@ -20,11 +20,12 @@ beforeAll(async () => {
   deposit = await prisma.deposit.create({
     data: {
       tracking_number: "53771ALBO11032024195558814",
-      status: CONFIG.SD.DEPOSIT_STATUS.PENDING,
-      player_id: player.id,
+      status: DEPOSIT_STATUS.PENDING,
+      Player: { connect: { id: player.id } },
       amount: 10,
       date: "2024-03-11T03:00:00.000Z",
       sending_bank: "90646",
+      CoinTransfer: { create: { status: COIN_TRANSFER_STATUS.PENDING } },
     },
   });
 });