+++ src/components/coin-transfers/controller.ts
@@ -5,16 +5,19 @@ import { apiResponse } from "@/helpers/apiResponse";
 export class CoinTransferControler {
   static async releasePending(_req: Req, res: Res, next: NextFn) {
     try {
-      // TODO
-      //   const { id } = req.params;
-      //   const { status } = req.body;
-      //   const coinTransfer = await CoinTransferServices.releasePending(
-      //     id,
-      //     status,
-      //   );
       const coinTransferServices = new CoinTransferServices();
       const coinTransfers = await coinTransferServices.releasePending();
       res.status(OK).json(apiResponse(coinTransfers));
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  static async getPendingTotal(_req: Req, res: Res, next: NextFn) {
+    try {
+      const coinTransferServices = new CoinTransferServices();
+      const amount = await coinTransferServices.getPendingTotal();
+      res.status(OK).json(apiResponse(amount));
     } catch (error) {
       next(error);
     }