+++ src/routes/bonus.router.ts
@@ -2,7 +2,11 @@ import { Router } from "express";
 import passport from "passport";
 import { checkExact } from "express-validator";
 import { BonusController } from "@/components/bonus/controller";
-import { requireAgentRole, requireUserRole } from "@/middlewares/auth";
+import {
+  requireAgentRole,
+  requireUserOrAgentRole,
+  requireUserRole,
+} from "@/middlewares/auth";
 import { validateResourceSearchRequest } from "@/components/players/validators";
 import { throwIfBadRequest } from "@/middlewares/requestErrorHandler";
 import {
@@ -26,6 +30,7 @@ bonusRouter.get(
 );
 bonusRouter.get(
   "/:id",
+  requireUserOrAgentRole,
   validateBonusId(),
   checkExact(),
   throwIfBadRequest,