+++ src/routes/transactions/deposits.router.ts
@@ -46,10 +46,6 @@ depositsRouter.get(
 depositsRouter.use(requireAgentRole);
 
 depositsRouter.get(
-  "/deposit/pending-coin-transfers",
-  DepositController.pendingCoinTransfers,
-);
-depositsRouter.get(
   "/deposit/:id",
   validateDepositIndex(),
   checkExact(),
@@ -57,8 +53,6 @@ depositsRouter.get(
   DepositController.show,
 );
 depositsRouter.post(
-  // TODO
-  // "/deposit/:id/update",
   "/deposit/:id/set-status",
   validateDepositIndex(),
   validateDepositSetStatusRequest(),